+++ lib/std/assert.ts
@@ -1,0 +1,4 @@
+export {
+  assertArrayIncludes,
+  assertEquals,
+} from "https://deno.land/std@0.202.0/assert/mod.ts";
