+++ lib/std/collections.ts
@@ -1,0 +1,2 @@
+export { distinct } from "https://deno.land/std@0.202.0/collections/distinct.ts";
+export { maxBy } from "https://deno.land/std@0.202.0/collections/max_by.ts";
