+++ lib/std/fs.ts
@@ -1,0 +1,1 @@
+export { existsSync } from "https://deno.land/std@0.202.0/fs/mod.ts";
