+++ lib/std/jsonc.ts
@@ -1,0 +1,1 @@
+export { parse } from "https://deno.land/std@0.202.0/jsonc/mod.ts";
