+++ lib/std/path.ts
@@ -1,0 +1,4 @@
+export { isAbsolute } from "https://deno.land/std@0.202.0/path/is_absolute.ts";
+export { relative } from "https://deno.land/std@0.202.0/path/relative.ts";
+export { resolve } from "https://deno.land/std@0.202.0/path/resolve.ts";
+export { toFileUrl } from "https://deno.land/std@0.202.0/path/to_file_url.ts";
