+++ lib/std/testing.ts
@@ -1,0 +1,9 @@
+export {
+  afterAll,
+  afterEach,
+  beforeAll,
+  beforeEach,
+  describe,
+  it,
+} from "https://deno.land/std@0.202.0/testing/bdd.ts";
+export { type Stub, stub } from "https://deno.land/std@0.202.0/testing/mock.ts";
