+++ lib/types.ts
@@ -1,0 +1,2 @@
+export type Maybe<T> = T | undefined;
+export type Brand<T, B> = T & { __brand: B };
