+++ lib/uri.ts
@@ -1,11 +1,6 @@
-import {
-  isAbsolute,
-  relative,
-  resolve,
-  toFileUrl,
-} from "https://deno.land/std@0.202.0/path/mod.ts";
+import { toFileUrl, isAbsolute, relative, resolve } from "./std/path.ts";
 
-type DefaultProtocol<Scheme extends string> = Scheme extends "http" | "https"
+export type DefaultProtocol<Scheme extends string> = Scheme extends "http" | "https"
   ? `${Scheme}://`
   : Scheme extends "file" ? `${Scheme}:///`
   : `${Scheme}:`;
