+++ lib/x/cliffy.ts
@@ -1,0 +1,3 @@
+export { Command } from "https://deno.land/x/cliffy@v1.0.0-rc.3/command/command.ts";
+export { colors } from "https://deno.land/x/cliffy@v1.0.0-rc.3/ansi/colors.ts";
+export { Select } from "https://deno.land/x/cliffy@v1.0.0-rc.3/prompt/select.ts";
