+++ lib/x/deno_graph.ts
@@ -1,0 +1,7 @@
+export {
+  createGraph,
+  type CreateGraphOptions,
+  init,
+  load,
+  type ModuleJson,
+} from "https://deno.land/x/deno_graph@0.55.0/mod.ts";
