+++ lib/x/import_map.ts
@@ -1,0 +1,5 @@
+export {
+  ImportMap,
+  type ImportMapJson,
+  parseFromJson,
+} from "https://deno.land/x/import_map@v0.15.0/mod.ts";
