+++ src/dependency.ts
@@ -1,0 +1,85 @@
+import type { Maybe } from "../lib/types.ts";
+import type { Path, SemVerString } from "./types.ts";
+import { parseSemVer } from "./semver.ts";
+
+export interface Dependency {
+  name: string;
+  version: SemVerString;
+  path: Path;
+}
+
+export const Dependency = {
+  parseProps,
+  resolveLatestSemVer,
+};
+
+function parseProps(
+  url: URL,
+): Maybe<Dependency> {
+  const body = url.hostname + url.pathname;
+  const semver = parseSemVer(url.href);
+  if (!semver) {
+    // The specifier does not contain a semver.
+    return;
+  }
+  const atSemver = "@" + semver;
+  const name = body.split(atSemver)[0];
+  const path = body.slice(name.length + atSemver.length);
+  return { name, version: semver, path: path as Path };
+}
+
+async function resolveLatestSemVer(
+  url: URL,
+): Promise<Maybe<SemVerString>> {
+  const props = parseProps(url);
+  if (!props) {
+    // The specifier is does not contain a semver.
+    return;
+  }
+  switch (url.protocol) {
+    case "npm:": {
+      const response = await fetch(
+        `https://registry.npmjs.org/${props.name}`,
+      );
+      if (!response.ok) {
+        throw new Error(`Failed to fetch npm registry: ${response.statusText}`);
+      }
+      const json = await response.json();
+      if (!json["dist-tags"]?.latest) {
+        throw new Error(
+          `Could not find the latest version of ${props.name} from registry.`,
+        );
+      }
+      const latestSemVer: string = json["dist-tags"].latest;
+      if (latestSemVer === props.version) {
+        // The dependency is up to date
+        return;
+      }
+      return latestSemVer as SemVerString;
+    }
+    case "http:":
+    case "https:": {
+      const response = await fetch(
+        url.protocol + "//" + props.name + props.path,
+        { method: "HEAD" },
+      );
+      await response.arrayBuffer();
+      if (!response.redirected) {
+        // The host did not redirect to a url with semver
+        return;
+      }
+      const specifierWithLatestSemVer = response.url;
+      if (specifierWithLatestSemVer === url.href) {
+        // The dependency is up to date
+        return;
+      }
+      return parseSemVer(specifierWithLatestSemVer);
+    }
+    case "node:":
+    case "file:":
+      return;
+    default:
+      // TODO: throw an error?
+      return;
+  }
+}
